+++ R/output.R
@@ -210,7 +210,7 @@ knit = function(input, output = NULL, tangle = FALSE, text = NULL,
     message('output file: ', normalizePath(output), ifelse(progress, '\n', ''))
   }
 
-  output %n% res
+  if (child_mode() || is.null(output)) res else output
 }
 #' @rdname knit
 #' @param documentation an integer specifying the level of documentation to go
@@ -307,50 +307,29 @@ auto_format = function(ext) {
 #' the result into the main document. It is designed to be used in the chunk
 #' option \code{child} and serves as the alternative to the
 #' \command{SweaveInput} command in Sweave.
-#'
-#' For LaTeX output, the command used to input the child document (usually
-#' \samp{input} or \samp{include}) is from the package option
-#' \code{child.command} (\code{opts_knit$get('child.command')}). For other types
-#' of output, the content of the compiled child document is returned.
-#'
-#' When we call \code{purl()} to extract R code, the code in the child document
-#' is extracted and saved into an R script.
-#'
-#' The path of the child document is relative to the parent document.
 #' @param ... arguments passed to \code{\link{knit}}
 #' @param eval logical: whether to evaluate the child document
-#' @return A character string of the form \samp{\command{child-doc.tex}} or
-#'   \code{source("child-doc.R")}, depending on the argument \code{tangle}
-#'   passed in. When concordance is turned on or the output format is not LaTeX,
-#'   the content of the compiled child document is returned as a character
-#'   string so it can be written back to the main document directly.
+#' @return A character string of the content of the compiled child document is
+#'   returned as a character string so it can be written back to the parent
+#'   document directly.
 #' @references \url{http://yihui.name/knitr/demo/child/}
 #' @note This function is not supposed be called directly like
 #'   \code{\link{knit}()}; instead it must be placed in a parent document to let
 #'   \code{\link{knit}()} call it indirectly.
+#'
+#'   The path of the child document is relative to the parent document.
 #' @export
-#' @examples ## you can write \Sexpr{knit_child('child-doc.Rnw')} in an Rnw file 'main.Rnw' to input child-doc.tex in main.tex
+#' @examples ## you can write \Sexpr{knit_child('child-doc.Rnw')} in an Rnw file 'main.Rnw' to input results from child-doc.Rnw in main.tex
 #'
 #' ## comment out the child doc by \Sexpr{knit_child('child-doc.Rnw', eval = FALSE)}
-#'
-#' ## use \include: opts_knit$set(child.command = 'include')
 knit_child = function(..., eval = TRUE) {
   if (!eval) return('')
   child = child_mode()
   opts_knit$set(child = TRUE) # yes, in child mode now
   on.exit(opts_knit$set(child = child)) # restore child status
-  path = knit(..., tangle = opts_knit$get('tangle'),
-              encoding = opts_knit$get('encoding') %n% getOption('encoding'))
-  if (is.null(path)) return() # the input document is empty
-  if (opts_knit$get('tangle')) {
-    str_c('\n', 'source("', path, '")')
-  } else if (concord_mode() || !out_format('latex')) {
-    on.exit(unlink(path)) # child output file is temporary
-    str_c(readLines(path), collapse = '\n')
-  } else {
-    path = gsub('[.]tex$', '', path, ignore.case = TRUE)
-    str_c('\n\\', opts_knit$get('child.command'), '{', path, '}')
-  }
+  res = knit(..., tangle = opts_knit$get('tangle'),
+             encoding = opts_knit$get('encoding') %n% getOption('encoding'))
+  paste(c('', res), collapse = '\n')
 }
 
 knit_log = new_defaults()  # knitr log for errors, warnings and messages